fix : 회원 슬라이스 테스트 - CRUD : 제목 수정, 오류 정정, 레거시 코드 삭제
